Fix bug - Limit to top 20 candidiates after DMR

+++ Backend/src/schema.py
@@ -54,7 +54,7 @@ class Metadata(GettableBaseModel):
     url: str
     viewportHeight: int
     viewportWidth: int
-    zoomLevel: int
+    zoomLevel: float
 
 
 class TransitionProperties(GettableBaseModel):
